Drum-Kit-Game
This is a playing drum game that contains various types of drums and other instruments also. The game is made with the proper frontend and along mp3 are included which will give the sound of the game when you will play it. The game is created using HTML, CSS, JS, BOOTSTRAP, and JQUERY.